# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  invelog_pkg_models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Container:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/invelog_pkg_models.Location'
      location_id:
        type: string
      name:
        type: string
      parent:
        $ref: '#/definitions/invelog_pkg_models.Container'
      parent_id:
        type: string
      project:
        $ref: '#/definitions/invelog_pkg_models.Project'
      project_id:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Item:
    properties:
      category:
        $ref: '#/definitions/invelog_pkg_models.Category'
      category_id:
        type: string
      checked_out:
        type: boolean
      container:
        $ref: '#/definitions/invelog_pkg_models.Container'
      container_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      individual_notes:
        type: string
      item_type:
        $ref: '#/definitions/invelog_pkg_models.ItemType'
      item_type_id:
        type: string
      last_check_out_time:
        type: string
      name:
        description: Used if ItemType is not provided
        type: string
      origin_location:
        $ref: '#/definitions/invelog_pkg_models.Location'
      origin_location_id:
        type: string
      quantity:
        type: integer
      serial_number:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.ItemType:
    properties:
      category:
        $ref: '#/definitions/invelog_pkg_models.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      manufacturer:
        type: string
      name:
        type: string
      part_number:
        type: string
      specifications:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Location:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  invelog_pkg_models.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        description: e.g., "active", "completed", "archived"
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive inventory management system backend for electronic
    components.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Invelog API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: Get all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/invelog_pkg_models.Project'
            type: array
      summary: List Projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: Project Data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/invelog_pkg_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Project
      tags:
      - Projects
  /projects/{id}:
    delete:
      description: Delete a project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Project
      tags:
      - Projects
    get:
      description: Get a project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Project'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update a project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project Data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/invelog_pkg_models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/invelog_pkg_models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Project
      tags:
      - Projects
  /search/items:
    get:
      description: Search across items using query parameters
      parameters:
      - description: Search query (matches name, description, notes)
        in: query
        name: query
        type: string
      - description: Filter by Category ID
        in: query
        name: category_id
        type: string
      - description: Filter by Container ID
        in: query
        name: container_id
        type: string
      - description: Filter by Project ID
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/invelog_pkg_models.Item'
            type: array
      summary: Search Items
      tags:
      - Search
swagger: "2.0"